# openapi.yaml
openapi: 3.1.0
info:
  title: PFF Fantasy League API
  version: "1.0.0"
  description: |
    REST interface for the PFF Fantasy League toolkit. All requests must include
    the header `X-API-Key` matching the `PFF_API_KEY` environment variable.
servers:
  - url: https://pff-fantasy-analyzer.onrender.com
    description: Production
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ConfigKnobs:
      type: object
      description: Scoring knobs currently in effect. Values are floats between 0 and 1 unless noted.
      additionalProperties: false
      properties:
        fuzzy_cutoff:
          type: number
          format: float
          description: Minimum similarity required when matching roster names.
        bench_ovar_beta:
          type: number
          format: float
          description: Weight applied to overall rank advantage when scoring bench players.
        combined_starters_weight:
          type: number
          format: float
          description: Share of the combined score attributed to starter strength.
        combined_bench_weight:
          type: number
          format: float
          description: Share of the combined score attributed to bench depth.
        projection_scale_beta:
          type: number
          format: float
          description: Influence of projection z-scores when rescaling rankings.
        bench_z_fallback_threshold:
          type: number
          format: float
          description: Threshold that swaps to percentile-based bench scoring when variance is low.
        bench_percentile_clamp:
          type: number
          format: float
          description: Clamp applied to percentile-based z-scores to limit outlier influence.
      required:
        - fuzzy_cutoff
        - bench_ovar_beta
        - combined_starters_weight
        - combined_bench_weight
        - projection_scale_beta
        - bench_z_fallback_threshold
        - bench_percentile_clamp
...
    LeaderboardEntry:
      type: object
      properties:
        team: {type: string}
        value: {type: number}
...
    TeamDetail:
      type: object
      properties:
        team: {type: string}
        starters:
          type: array
          items: {$ref: '#/components/schemas/StarterDetail'}
        bench:
          type: array
          items: {$ref: '#/components/schemas/BenchDetail'}
        benchLimit: {type: integer, nullable: true}
    TeamEvaluation:
      type: object
      properties:
        team: {type: string}
        combinedScore: {type: number}
        starterVOR: {type: number}
        benchScore: {type: number}
        starterProjection: {type: number}
    EvaluateRequest:
      type: object
      properties:
        projectionScaleBeta: {type: number, nullable: true}
        replacementSkipPct: {type: number, nullable: true}
        replacementWindow: {type: integer, nullable: true}
        benchOvarBeta: {type: number, nullable: true}
        combinedStartersWeight: {type: number, nullable: true}
        combinedBenchWeight: {type: number, nullable: true}
        benchZFallbackThreshold: {type: number, nullable: true}
        benchPercentileClamp: {type: number, nullable: true}
        scarcitySampleStep: {type: number, nullable: true}
        rankingsPath: {type: string, nullable: true}
        projectionsPath: {type: string, nullable: true}
        supplementalPath: {type: string, nullable: true}
        rosters:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items: {type: string}
        includeDetails: {type: boolean, default: false}
        benchLimit: {type: integer, nullable: true}
    EvaluateResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        playerCount: {type: integer}
        teams:
          type: array
          items: {$ref: '#/components/schemas/TeamEvaluation'}
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
        settings:
          $ref: '#/components/schemas/ConfigKnobs'
        replacementPoints:
          type: object
          additionalProperties: {type: number}
        replacementTargets:
          type: object
          additionalProperties: {type: number}
        scarcitySamples:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                slot: {type: number}
                projection: {type: number}
        zeroSum:
          $ref: '#/components/schemas/ZeroSumResponse'
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
...
    ZeroSumCombinedGroup:
      allOf:
        - $ref: '#/components/schemas/ZeroSumGroup'
        - type: object
          properties:
            weights:
              type: object
              additionalProperties: {type: number}
          required: [weights]
...
    ZeroSumResponse:
      type: object
      properties:
        teamCount: {type: integer}
        starters: {$ref: '#/components/schemas/ZeroSumGroup'}
        bench: {$ref: '#/components/schemas/ZeroSumGroup'}
        combined: {$ref: '#/components/schemas/ZeroSumCombinedGroup'}
        positions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        benchPositions:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        slots:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumGroup'}
        flex:
          $ref: '#/components/schemas/ZeroSumGroup'
          nullable: true
        analytics:
          $ref: '#/components/schemas/ZeroSumAnalytics'
      required: [teamCount, starters, bench, combined, positions, benchPositions, slots, analytics]
...
    ZeroSumShift:
      type: object
      properties:
        combined:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        starters:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        bench:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        positions:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        benchPositions:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        slots:
          type: object
          additionalProperties:
            type: object
            additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        flex:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ZeroSumShiftMetrics'}
        summary: {type: object}
      required: [combined, starters, bench, positions, benchPositions, slots, flex, summary]
    EvaluateDeltaSnapshot:
      type: object
      properties:
        teams:
          type: array
          items: {$ref: '#/components/schemas/TeamEvaluation'}
        zeroSum:
          $ref: '#/components/schemas/ZeroSumResponse'
      required: [teams, zeroSum]
    EvaluateDeltaResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        baseline: {$ref: '#/components/schemas/EvaluateDeltaSnapshot'}
        scenario: {$ref: '#/components/schemas/EvaluateDeltaSnapshot'}
        zeroSumShift: {$ref: '#/components/schemas/ZeroSumShift'}
      required: [evaluatedAt, baseline, scenario, zeroSumShift]
...
    TradePiece:
      type: object
      properties:
        group: {type: string}
        name: {type: string}
      required: [group, name]
    TradeTeamResult:
      type: object
      properties:
        team: {type: string}
        baseline: {type: number}
        postTrade: {type: number}
        delta: {type: number}
    TradeEvaluateRequest:
      type: object
      properties:
        teamA: {type: string}
        teamB: {type: string}
        sendA:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        sendB:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        includeDetails: {type: boolean, default: true}
        benchLimit: {type: integer, nullable: true}
      required: [teamA, teamB]
    TradeEvaluateResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        teams:
          type: array
          items: {$ref: '#/components/schemas/TradeTeamResult'}
        combinedScores:
          type: object
          additionalProperties: {type: number}
        replacementPoints:
          type: object
          additionalProperties: {type: number}
        replacementTargets:
          type: object
          additionalProperties: {type: number}
        starterVOR:
          type: object
          additionalProperties: {type: number}
        benchTotals:
          type: object
          additionalProperties: {type: number}
        zeroSumBefore:
          $ref: '#/components/schemas/ZeroSumResponse'
        zeroSumAfter:
          $ref: '#/components/schemas/ZeroSumResponse'
        zeroSumShift:
          $ref: '#/components/schemas/ZeroSumShift'
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
    TradeProposal:
      type: object
      properties:
        sendA:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        sendB:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        receiveA:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        receiveB:
          type: array
          items: {$ref: '#/components/schemas/TradePiece'}
        combinedScores:
          type: object
          additionalProperties: {type: number}
        delta:
          type: object
          additionalProperties: {type: number}
        score: {type: number}
        acceptance: {type: number}
        fairnessSplit: {type: number, nullable: true}
        dropTax:
          type: object
          additionalProperties: {type: number}
        starGain:
          type: object
          additionalProperties: {type: number}
        narrative:
          type: object
          additionalProperties: {type: string}
        details:
          type: array
          items: {$ref: '#/components/schemas/TeamDetail'}
          nullable: true
        leaderboards:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/LeaderboardEntry'}
          nullable: true
    TradeFindRequest:
      type: object
      properties:
        teamA: {type: string}
        teamB: {type: string}
        maxPlayers: {type: integer, default: 3}
        playerPool: {type: integer, default: 15}
        topResults: {type: integer, default: 5}
        topBench: {type: integer, default: 5}
        minGainA: {type: number, default: 0.0}
        maxLossB: {type: number, default: 0.25}
        pruneMargin: {type: number, default: 0.05}
        minUpperBound: {type: number, default: -5.0}
        fairnessMode: {type: string, default: sum, enum: [sum, weighted, nash]}
        fairnessSelfBias: {type: number, default: 0.6}
        fairnessPenaltyWeight: {type: number, default: 0.5}
        consolidationBonus: {type: number, default: 0.0}
        dropTaxFactor: {type: number, default: 0.5}
        acceptanceFairnessWeight: {type: number, default: 0.4}
        acceptanceNeedWeight: {type: number, default: 0.35}
        acceptanceStarWeight: {type: number, default: 0.25}
        acceptanceNeedScale: {type: number, default: 1.0}
        starVorScale: {type: number, default: 60.0}
        dropTaxAcceptanceWeight: {type: number, default: 0.02}
        narrativeOn: {type: boolean, default: true}
        minAcceptance: {type: number, default: 0.2}
        mustSendA:
          type: array
          items: {type: string}
        mustReceiveB:
          type: array
          items: {type: string}
        includeDetails: {type: boolean, default: true}
        benchLimit: {type: integer, nullable: true}
      required: [teamA, teamB]
    TradeFindResponse:
      type: object
      properties:
        evaluatedAt: {type: string, format: date-time}
        baselineCombined:
          type: object
          additionalProperties: {type: number}
        proposals:
          type: array
          items: {$ref: '#/components/schemas/TradeProposal'}
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed]
    JobInfo:
      type: object
      properties:
        jobId: {type: string}
        jobType: {type: string}
        status: {$ref: '#/components/schemas/JobStatus'}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        finishedAt: {type: string, format: date-time, nullable: true}
        result: {type: object, nullable: true}
        error: {type: string, nullable: true}
        metadata: {type: object}
      required: [jobId, jobType, status, createdAt, metadata]
    JobCreatedResponse:
      type: object
      properties:
        jobId: {type: string}
        jobType: {type: string}
        status: {$ref: '#/components/schemas/JobStatus'}
        pollUrl: {type: string}
      required: [jobId, jobType, status, pollUrl]
paths:
...
  /evaluate:
    post:
      tags: [league]
      summary: Evaluate league
      operationId: evaluateLeague
      description: |
        Submit an `EvaluateRequest` JSON body to tweak knobs and hypothetical rosters. For convenience,
        `includeDetails` and `benchLimit` may also be supplied as query parameters when you only need
        to toggle those flags.
      x-openai-isConsequential: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRequest'
            examples:
              default:
                summary: Baseline evaluation
                value: {}
              withDetails:
                summary: Include details with custom bench limit
                value:
                  includeDetails: true
                  benchLimit: 3
      responses:
        '200':
          description: Evaluation result
          content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateResponse'
  /evaluate/delta:
    post:
      tags: [league]
      summary: Compare baseline and modified rosters
      operationId: evaluateLeagueDelta
      description: |
        Run two evaluations (baseline vs. hypothetical roster override) and return both results alongside
        a zero-sum shift summary (share/surplus/value deltas with winners/losers callouts).
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRequest'
      responses:
        '200':
          description: Baseline and scenario comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateDeltaResponse'
...
  /trade/evaluate:
    post:
      tags: [trade]
      summary: Evaluate trade proposal
      operationId: evaluateTrade
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeEvaluateRequest'
            example:
              teamA: JohnBurrows_School
              teamB: KWood_Super_Marios
              sendA:
                - {group: WR, name: George Pickens}
              sendB:
                - {group: WR, name: Drake London}
              includeDetails: true
      responses:
        '200':
          description: Trade impact summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeEvaluateResponse'
  /trade/find:
    post:
      tags: [trade]
      summary: Search for favorable trades
      operationId: findTrades
      description: |
        Submit a JSON body described by `TradeFindRequest`. For quick ad-hoc usage you may also pass
        the same fields as query parameters (teamA, teamB, maxPlayers, playerPool, topResults,
        topBench, includeDetails, benchLimit, runAsync), but the canonical contract uses the JSON body.
      x-openai-isConsequential: false
      requestBody:
        required: false
        description: Use a JSON payload when you want to control multiple search knobs (fairness weights, narrative toggles, etc.).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeFindRequest'
            example:
              teamA: JohnBurrows_School
              teamB: KWood_Super_Marios
              maxPlayers: 3
              playerPool: 24
              topResults: 10
              includeDetails: true
      responses:
        '200':
          description: Candidate trades ordered by objective score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeFindResponse'
        '202':
          description: Search accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'
        '400':
          description: Invalid input (e.g., unknown team or bad knob value).
        '422':
          description: Validation error when the JSON body is malformed.
...
  /jobs/{jobId}:
    get:
      tags: [jobs]
      summary: Fetch background job status
      operationId: getJob
      x-openai-isConsequential: false
      parameters:
        - name: jobId
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobInfo'
        '404':
          description: Job not found
security:
  - ApiKeyAuth: []